+++ genomepy/cli.py
@@ -135,17 +135,16 @@ def get_install_options():
     add provider in front of the provider specific options to prevent overlap"""
     install_options = general_install_options
 
-    # only run if -h or --help are in the arguments
-    if "-h" in click.get_os_args() or "--help" in click.get_os_args():
-        for name in genomepy.ProviderBase.list_providers():
-            # directly access the provider's arguments, skips __init__
+    # extend install options with provider specific options
+    if "install" in click.get_os_args():
+        for provider in genomepy.ProviderBase.list_providers():
             p_dict = eval(
                 "genomepy.provider."
-                + name.capitalize()
-                + "Provider.list_install_options(None)"
+                + provider.capitalize()
+                + "Provider.provider_specific_install_options"
             )
             for option in p_dict.keys():
-                p_dict[option]["long"] = name + "-" + p_dict[option]["long"]
+                p_dict[option]["long"] = provider + "-" + p_dict[option]["long"]
             install_options.update(p_dict)
 
     return install_options