+++ tests/04_genome_test.py
@@ -266,6 +266,7 @@ def test_get_track_type():
         (["chr1:10-20", "chr2:10-20"], "interval"),
         ("tests/data/regions.txt", "interval"),
         ("tests/data/regions.bed", "bed"),
+        ("tests/data/regions2.bed", "bed"),
     ]
 
     for track, track_type in tracks:
@@ -300,13 +301,29 @@ def test_track2fasta(genome="tests/data/small_genome.fa.gz"):
 
 def test_sizes(genome="tests/data/gap.fa"):
     g = genomepy.Genome(genome)
-    g.contig_sizes()
+    assert list(g.sizes.keys()) == ["chr1", "chr2", "chr3"]
+    assert all(isinstance(g.sizes[chrom], int) for chrom in g.sizes.keys())
+    assert g.sizes["chr1"] == 28
+
+    # does not overwrite user-set sizes
+    g.sizes = {"asd": 1}
+    assert g.sizes == {"asd": 1}
+
+    # repopulates empty dicts
+    g.sizes = {}
     assert list(g.sizes.keys()) == ["chr1", "chr2", "chr3"]
 
 
 def test_gaps(genome="tests/data/gap.fa"):
     g = genomepy.Genome(genome)
-    g.gap_sizes()
+    assert list(g.gaps.keys()) == ["chr1", "chr3"]
+
+    # does not overwrite user-set gaps
+    g.gaps = {"asd": 1}
+    assert g.gaps == {"asd": 1}
+
+    # repopulates empty dicts
+    g.gaps = {}
     assert list(g.gaps.keys()) == ["chr1", "chr3"]